move navbar components to structure/ directory
