+++ components/src/icons/icon-data.js
@@ -11,8 +11,8 @@ export const WIFI = 'wifi'
 export const FLASK = 'flask'
 export const CHECKED = 'checked'
 export const UNCHECKED = 'unchecked'
-export const EXPAND = 'expand'
-export const COLLAPSE = 'collapse'
+export const CHEVRON_UP = 'chevron up'
+export const CHEVRON_DOWN = 'chevron down'
 export const PROTOCOL = 'protocol'
 export const COG = 'cog'
 export const CONNECT = 'connect'
@@ -55,11 +55,11 @@ const ICON_DATA_BY_NAME = {
     viewBox: '0 0 24 24',
     path: 'M12,21.6 C6.6980664,21.6 2.4,17.3019336 2.4,12 C2.4,6.6980664 6.6980664,2.4 12,2.4 C14.5460783,2.4 16.9878759,3.41142567 18.7882251,5.2117749 C20.5885743,7.01212413 21.6,9.4539217 21.6,12 C21.6,17.3019336 17.3019336,21.6 12,21.6 Z M12,5.32907052e-16 C5.372583,-2.13162821e-15 5.32907052e-16,5.372583 0,12 C-1.59872116e-15,15.1825979 1.26428208,18.2348448 3.51471863,20.4852814 C5.76515517,22.7357179 8.81740212,24 12,24 C15.1825979,24 18.2348448,22.7357179 20.4852814,20.4852814 C22.7357179,18.2348448 24,15.1825979 24,12 C24,8.81740212 22.7357179,5.76515517 20.4852814,3.51471863 C18.2348448,1.26428208 15.1825979,2.66453526e-16 12,5.32907052e-16 Z'
   },
-  [EXPAND]: {
+  [CHEVRON_UP]: {
     viewBox: '0 0 24 24',
     path: 'M7.41 9L12 13.242 16.59 9 18 10.312l-6 5.545-6-5.545z'
   },
-  [COLLAPSE]: {
+  [CHEVRON_DOWN]: {
     viewBox: '0 0 24 24',
     path: 'M7.41 15.857L12 11.615l4.59 4.242L18 14.545 12 9l-6 5.545z'
   },
