+++ components/src/index.js
@@ -20,4 +20,3 @@ export * from './CenteredTextSvg'
 export * from './icons'
 export * from './structure'
 export * from './lists'
-export * from './navbar'
