--- components/src/navbar/index.js
@@ -1,6 +1,0 @@
-// @flow
-// NavBar container and buttons
-import VerticalNavBar from './vertical-navbar'
-import NavButton from './nav-button'
-
-export {VerticalNavBar, NavButton}
