+++ components/src/structure/index.js
@@ -2,5 +2,7 @@
 // structure components
 import PageTabs from './PageTabs'
 import TitleBar from './TitleBar'
+import VerticalNavBar from './vertical-navbar'
+import NavButton from './nav-button'
 
-export {PageTabs, TitleBar}
+export {PageTabs, TitleBar, VerticalNavBar, NavButton}
