+++ components/src/structure/vertical-navbar.js
